Crawler deletes temporary files after they are stored
But due to the cyclic port connections the master cannot shutdown the
topology cleanly. But this is ok for now as the state is saved
elsewhere (mongodb, redis, files etc.).

In the future, we may either need to change delete_port to a query port
or figure out a more structured way to handle tokens on temporary
files.